move control transforms out of load
